refactor: 각 Enum에 대한 테스트를 타입별로 분리
